+++ src/main/java/net/geforcemods/securitycraft/network/server/AssembleBlockPocket.java
@@ -2,9 +2,15 @@ package net.geforcemods.securitycraft.network.server;
 
 import java.util.function.Supplier;
 
+import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.network.NetworkEvent;
 
 public class AssembleBlockPocket {
@@ -29,10 +35,17 @@ public class AssembleBlockPocket {
 	}
 
 	public void handle(Supplier<NetworkEvent.Context> ctx) {
-		if (ctx.get().getSender().level.getBlockEntity(pos) instanceof BlockPocketManagerBlockEntity be && be.isOwnedBy(ctx.get().getSender())) {
+		Player player = ctx.get().getSender();
+
+		if (player.level.getBlockEntity(pos) instanceof BlockPocketManagerBlockEntity be && be.isOwnedBy(player)) {
+			MutableComponent feedback;
+
 			be.setSize(size);
-			be.autoAssembleMultiblock();
+			feedback = be.autoAssembleMultiblock();
 			be.setChanged();
+
+			if (feedback != null)
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, ChatFormatting.DARK_AQUA);
 		}
 	}
 }