+++ src/main/java/net/geforcemods/securitycraft/network/server/ToggleBlockPocketManager.java
@@ -2,9 +2,16 @@ package net.geforcemods.securitycraft.network.server;
 
 import java.util.function.Supplier;
 
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
+import net.geforcemods.securitycraft.network.client.BlockPocketManagerFailedActivation;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.network.NetworkEvent;
 
@@ -37,12 +44,21 @@ public class ToggleBlockPocketManager {
 		Player player = ctx.get().getSender();
 
 		if (player.level.getBlockEntity(pos) instanceof BlockPocketManagerBlockEntity be && be.isOwnedBy(player)) {
+			MutableComponent feedback;
+
 			be.setSize(size);
 
 			if (enabling)
-				be.enableMultiblock();
+				feedback = be.enableMultiblock();
 			else
-				be.disableMultiblock();
+				feedback = be.disableMultiblock();
+
+			if (feedback != null) {
+				if (enabling && !be.isEnabled())
+					SecurityCraft.CHANNEL.reply(new BlockPocketManagerFailedActivation(pos), ctx.get());
+
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, ChatFormatting.DARK_AQUA, false);
+			}
 
 			be.setChanged();
 		}